Change produitId type and remove price column from list
Updated VenteProduitStatDTO to use Long for produitId instead of Integer for consistency with entity IDs. Removed the 'Prix de vente' column from the produit list template to simplify the UI.

+++ src/main/java/com/gestioncafe/dto/VenteProduitStatDTO.java
@@ -3,19 +3,19 @@ package com.gestioncafe.dto;
 import java.math.BigDecimal;
 
 public class VenteProduitStatDTO {
-    private final Integer produitId;
+    private final Long produitId;
     private final String produitNom;
     private final BigDecimal quantiteTotale;
     private final BigDecimal beneficeTotal;
 
-    public VenteProduitStatDTO(Integer produitId, String produitNom, BigDecimal quantiteTotale, BigDecimal beneficeTotal) {
+    public VenteProduitStatDTO(Long produitId, String produitNom, BigDecimal quantiteTotale, BigDecimal beneficeTotal) {
         this.produitId = produitId;
         this.produitNom = produitNom;
         this.quantiteTotale = quantiteTotale;
         this.beneficeTotal = beneficeTotal;
     }
 
-    public Integer getProduitId() {
+    public Long getProduitId() {
         return produitId;
     }
